test: fix and share common code for checkout tests
